move Context into its own module
